+++ BE/house/src/main/java/com/nextsquad/house/login/oauth/GithubOauthClient.java
@@ -5,19 +5,35 @@ import com.nextsquad.house.dto.login.GithubAccessTokenResponseDto;
 import com.nextsquad.house.login.userinfo.UserInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
 import java.util.Map;
 
 @Slf4j
-public class GithubOauthClient extends OauthClient {
+public class GithubOauthClient implements OauthClient {
+
+    private final String clientId;
+    private final String authServerUrl;
+    private final String resourceServerUrl;
+    private final String secretKey;
+    private final WebClient webClient = WebClient.create();
+
     public GithubOauthClient(String clientId, String authServerUrl, String resourceServerUrl, String secretKey) {
-        super(clientId, authServerUrl, resourceServerUrl, secretKey);
+        this.clientId = clientId;
+        this.authServerUrl = authServerUrl;
+        this.resourceServerUrl = resourceServerUrl;
+        this.secretKey = secretKey;
     }
 
     @Override
-    protected String getAccessToken(String authCode) {
+    public UserInfo getUserInfo(String authCode) {
+        String accessToken = getAccessToken(authCode);
+        return getOauthUserInfo(accessToken);
+    }
+
+    private String getAccessToken(String authCode) {
         GithubAccessTokenResponseDto rawToken = webClient.post()
                 .uri(authServerUrl, uriBuilder -> uriBuilder
                         .queryParam("client_id", clientId)
@@ -36,8 +52,7 @@ public class GithubOauthClient extends OauthClient {
         return parseToken(rawToken.getAccessToken());
     }
 
-    @Override
-    protected UserInfo getOauthUserInfo(String accessToken) {
+    private UserInfo getOauthUserInfo(String accessToken) {
         String response = webClient.get()
                 .uri(resourceServerUrl)
                 .header("authorization", accessToken)
@@ -50,12 +65,11 @@ public class GithubOauthClient extends OauthClient {
         return convertToUserInfoFrom(response);
     }
 
-    @Override
-    protected String parseToken(String rawToken) {
+    private String parseToken(String rawToken) {
         return String.format("token %s", rawToken);
     }
 
-    protected UserInfo convertToUserInfoFrom(String rawInfo) {
+    private UserInfo convertToUserInfoFrom(String rawInfo) {
         Map<String, String> infoMap = new Gson().fromJson(rawInfo, Map.class);
         return new UserInfo(infoMap.get("login"), infoMap.get("login"), null, OauthClientType.GITHUB);
     }
