+++ BE/house/src/main/java/com/nextsquad/house/login/oauth/NaverOauthClient.java
@@ -6,22 +6,36 @@ import com.nextsquad.house.login.userinfo.UserInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
 
 @Slf4j
-public class NaverOauthClient extends OauthClient{
+public class NaverOauthClient implements OauthClient {
+    private final String clientId;
+    private final String authServerUrl;
+    private final String resourceServerUrl;
+    private final String secretKey;
+    private final WebClient webClient = WebClient.create();
 
     @Value("${spring.oauth.naver.state}")
     private String state;
 
     public NaverOauthClient(String clientId, String authServerUrl, String resourceServerUrl, String secretKey) {
-        super(clientId, authServerUrl, resourceServerUrl, secretKey);
+        this.clientId = clientId;
+        this.authServerUrl = authServerUrl;
+        this.resourceServerUrl = resourceServerUrl;
+        this.secretKey = secretKey;
     }
 
     @Override
-    protected String getAccessToken(String authCode) {
+    public UserInfo getUserInfo(String authCode) {
+        String accessToken = getAccessToken(authCode);
+        return getOauthUserInfo(accessToken);
+    }
+
+    private String getAccessToken(String authCode) {
         NaverAccessTokenResponseDto rawToken = webClient.post()
                 .uri(authServerUrl, uriBuilder -> uriBuilder
                         .queryParam("grant_type", "authorization_code")
@@ -42,8 +56,7 @@ public class NaverOauthClient extends OauthClient{
         return parseToken(rawToken.getAccessToken());
     }
 
-    @Override
-    protected UserInfo getOauthUserInfo(String accessToken) {
+    private UserInfo getOauthUserInfo(String accessToken) {
         NaverUserInfoDto infoDto = webClient.get()
                 .uri(resourceServerUrl)
                 .header("authorization", accessToken)
@@ -56,8 +69,7 @@ public class NaverOauthClient extends OauthClient{
         return infoDto.toUserInfo();
     }
 
-    @Override
-    protected String parseToken(String rawToken) {
+    private String parseToken(String rawToken) {
         return String.format("Bearer %s", rawToken);
     }
